+++ CRMAPI/CRMAPI/Models/Produto.Model.cs
@@ -9,7 +9,7 @@ namespace CRMAPI.Models
         public string NomeProduto { get; set; }
         public int IdProduto { get; set; }
         public int IdCriador { get; set; }
-        public double? ValorMinimo { get; set; }
+        public decimal? ValorMinimo { get; set; }
         public DateTime DataCriacao { get; set; }
         public DateTime? DataAlteracao { get; set; }
         public DateTime? DataDesativado { get; set; }
